+++ src/Components/EventMain/ Recommend.scss
@@ -2,10 +2,7 @@
 
 .EventMain {
   .rec {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
+    @include center;
     h4 {
       padding: 80px 0 36px;
       text-align: center;
