+++ src/Components/OrderList/MyOrderList/List.scss
@@ -1,4 +1,4 @@
-@import "../../Styles/common";
+@import "../../../Styles/common";
 
 .OrderList {
   .product {
