+++ src/Components/OrderList/MyOrderList/MyOrderList.scss
@@ -1,4 +1,4 @@
-@import "../../Styles/common";
+@import "../../../Styles/common";
 
 .OrderList {
   .list-detail {
