+++ src/Styles/common.scss
@@ -45,3 +45,10 @@ input[type="checkbox"] {
 @mixin none {
   display: none;
 }
+
+@mixin center {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
